driver: cache stats values
This PR ensures a task's cgroup will only be scanned once
every 10 seconds, instead of being driven to scan once per
second by requests from Nomad.

This caching strategy was also part of Nomad <= 1.6 but got
removed in 1.7; we need to add it back for the builtin drivers
as well.