driver: keep shim alive to report exit code
This PR fixes a design flaw where the exit code of a task process would
not be retrievable after a Nomad client restart. This is critical for
batch jobs, the success or failure of which are determined by their exit
code.

The fix is basically to have the exec2-shim process remain alive as a
parent of the task process and record the exit code to a known file
inside the task directory. That file is then read back by the plugin
after the task process exits. This path is only necessary in the reattach
case; wait.go now splits logic between waiting on a child and waiting
on an orphan. The orphan case makes use of the recent pidfd_open syscall
to acquire an FD on which we call poll, which returns when the process
exits. Unfortunately Linux does not yet provide a way to retrieve the
exit status information via the pidfd_* set of syscalls, and so we fallback
to the shim + file hack.

While fixing this I also realized the FIFO filepaths for logging were
incorrect - we were using the host-perspective paths instead of the
task mounts directory perspective. This is a requirement for working in
a production - hardened environment, and the bug became apparent because
the tests would no longer pass without the fix. The reason is because
before the FIFOs were opened by the plugin (running as root) and the FD's
simply passed through unshare upon launch. Now, the shim opens the FIFOs
directly and would fail to do so with the incorrect host-perspective paths
due to them being owned by root.