exec2: do not allow /tmp as part of the defaults
This PR removes the unveil-ing of /tmp at rwc permission level as part
of the `unveil_defaults` plugin configuration.

A task sandbox already has a tmp directory created and is pointed to by
the standard TMPDIR environment variable; compliant tools should already
be using it.

Disallowing /tmp by default improves two things
- /tmp is readable and writable by all users (i.e. 777) and is an easy
place for other processes to leak information.
- /tmp is not accounted for in the task sandbox disk quota and may even
be tmpfs, wasting memory if a task is writing lots of temp data

A task that wants to make use of the real /tmp can do so if the operator
adds /tmp to the plugin unveil_paths configuration, or if unveil_by_task
is set in plugin configuration and the task adds /tmp to its unveil paths
list.