config: unveil NOMAD_* paths if unveil_defaults is set
This PR changes the driver such that NOMAD_TASK_DIR, NOMAD_ALLOC_DIR, and
NOMAD_SECRETS_DIR are automatically unveil'd with rwxc permissions if
the plugin `config.unveil_defaults` value is set. Previously we would do
this if and only if the task's `config.unveil` value was empty, which does
not make sense.

Also re-enable test cases that were disabled when we fixed the /tmp
unveil behavior. The unit tests had to be tweaked to actually resolve
the environment variables at run time because the testing harness does
not do that and figuring out how to make it work was too hard.