packaging: fix plugins directory filepath
When installing Nomad via Linux packaging the data directory is placed
at /opt/nomad/data, and the default plugins directory is a child of
that, i.e. /opt/nomad/data/plugins.

This PR fixes the linux package for exec2 to symlink the driver into the
correct path (it left off the /data/ part) such that the driver works
out of the box when doing:

apt/yum install nomad nomad-driver-exec2