driver: capture unshare/nsenter errors in task log
Fixes #20

On errors unshare and nsenter may contain output. This is a proof of
concept of sending that output to the task logs.

This also plumbs the driver's logger into the shim environment which
seems to work fine and helped immensely with debugging.

Not ready for merging. Needs tests and cleaning up.